// frontend/src/types/collection/FolderItem.ts
import type { RequestItem } from "../request/RequestItem";

export interface FolderItem {
  id: string;
  name: string;
  collectionId: string;
  parentFolderId?: string | null;
  subfolderCount: number;
  subfolders?: FolderItem[];
  requests?: RequestItem[];
  isLoaded?: boolean;
  isLoading?: boolean;
}